requirements.txt
fastapi==0.115.0
httpx==0.27.0
pydantic==2.8.0
pydantic-settings==2.5.0
aiofiles==24.1.0
uvicorn==0.30.0
python-multipart==0.0.9
pyyaml==6.0.1
langchain-openai==0.2.0
langgraph==0.2.0